+++ src/api/endpoints/admin.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends
 
 from src.api.schemas import AdminUserRequest
 from src.api.services.admin_service import AdminService
-from src.depends import Container
+from src.core.depends.container import Container
 
 admin_user_router = APIRouter()
