+++ src/api/endpoints/categories.py
@@ -5,7 +5,7 @@ from src.api.auth import check_header_contains_token
 from src.api.schemas import CategoryRequest, CategoryResponse
 from src.api.services import CategoryService
 from src.core.db.models import Category
-from src.depends import Container
+from src.core.depends.container import Container
 
 category_router = APIRouter(dependencies=[Depends(check_header_contains_token)])
