+++ src/api/endpoints/external_site_user.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends
 from src.api.auth import check_header_contains_token
 from src.api.schemas import ExternalSiteUserRequest
 from src.api.services import ExternalSiteUserService
-from src.depends import Container
+from src.core.depends.container import Container
 
 site_user_router = APIRouter(dependencies=[Depends(check_header_contains_token)])
