+++ src/api/endpoints/notification.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends
 from src.api.auth import check_header_contains_token
 from src.api.schemas import InfoRate, MessageList, TelegramNotificationRequest, TelegramNotificationUsersRequest
 from src.api.services.messages import TelegramNotificationService
-from src.depends import Container
+from src.core.depends.container import Container
 
 notification_router = APIRouter(dependencies=[Depends(check_header_contains_token)])
 log = structlog.get_logger()
