+++ src/api/events.py
@@ -3,8 +3,8 @@ from fastapi import FastAPI
 from telegram.ext import Application
 
 from src.bot import shutdown_bot, startup_bot
+from src.core.depends.container import Container
 from src.core.utils import set_ngrok
-from src.depends import Container
 from src.settings import Settings
