+++ src/bot/handlers/menu.py
@@ -9,8 +9,8 @@ from src.bot.keyboards import get_back_menu, get_menu_keyboard, get_no_mailing_k
 from src.bot.services.unsubscribe_reason import UnsubscribeReasonService
 from src.bot.services.user import UserService
 from src.bot.utils import delete_previous_message
+from src.core.depends.container import Container
 from src.core.logging.utils import logger_decor
-from src.depends import Container
 
 log = structlog.get_logger()
